refactor: move shap plot into new module
